+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/runtime/GCNode.java
@@ -1,7 +1,6 @@
 package org.enso.interpreter.node.expression.builtin.runtime;
 
 import com.oracle.truffle.api.CompilerDirectives;
-import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.nodes.Node;
 import org.enso.interpreter.dsl.BuiltinMethod;
 import org.enso.interpreter.runtime.EnsoContext;
@@ -9,27 +8,25 @@ import org.enso.interpreter.runtime.EnsoContext;
 @BuiltinMethod(
     type = "Runtime",
     name = "gc",
-    description = "Forces garbage collection",
+    description = "Forces garbage collection & other cleanups",
     autoRegister = false)
-public abstract class GCNode extends Node {
+public final class GCNode extends Node {
+  private GCNode() {}
 
-  public abstract Object execute();
+  @CompilerDirectives.TruffleBoundary
+  public Object execute(boolean flushCaches) {
+    var ctx = EnsoContext.get(this);
+    if (flushCaches) {
+      ctx.getResourceManager().scheduleFinalizationOfSystemReferences();
+    }
+    System.gc();
+    return ctx.getBuiltins().nothing();
+  }
 
   /**
    * @return A new GCNode.
    */
-  public static GCNode build() {
-    return GCNodeGen.create();
-  }
-
-  @Specialization
-  Object doGc() {
-    runGC();
-    return EnsoContext.get(this).getBuiltins().nothing();
-  }
-
-  @CompilerDirectives.TruffleBoundary
-  private void runGC() {
-    System.gc();
+  static GCNode build() {
+    return new GCNode();
   }
 }