+++ CHANGELOG.md
@@ -15,10 +15,12 @@
 - [`Panic.rethrow` keeps original location][14480]
 - [Use `State.get if_missing` to avoid too frequent `Panic.throw`][14490]
 - [Lazily initialized local variables with Ref.memoize][14554].
+- [Flush system caches via Runtime.gc][14557]
 
 [14480]: https://github.com/enso-org/enso/pull/14480
 [14490]: https://github.com/enso-org/enso/pull/14490
 [14536]: https://github.com/enso-org/enso/pull/14554
+[14557]: https://github.com/enso-org/enso/pull/14557
 
 # Next Release